plant/app/requirements.txt
pyzmq==25.1.1
pyyaml==6.0.1
numpy==1.24.3